fix(experiments): renaming files to match other description components.
