+++ frontend/src/scenes/experiments/activity-descriptions/metricChangeDescriptions.tsx
